Fix portability bug in init snippet
The problem with this line
```
[[ -s "${SDKMAN_DIR}/bin/sdkman-init.sh" ]] && source "${SDKMAN_DIR}/bin/sdkman-init.sh"
```
is, that it breaks the portability of any `.bashrc` or `.zshrc` in the case, that one uses _different_ users with the same set of dotfiles.

With the fix proposed here, we get rid of this portability problem. Just search on github for `THIS MUST BE AT THE END OF THE FILE FOR SDKMAN TO WORK` to see how many people hardcoded their dotfiles to work only with a specific $HOME/user name. The moment the user name changes it breaks their sdkman installation.